requirements_new.txt
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
bcrypt==4.3.0
binaryornot==0.4.4
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cookiecutter==2.6.0
Faker==37.4.0
fastapi==0.115.13
flet==0.28.3
flet-cli==0.28.3
flet-desktop==0.28.3
flet-web==0.28.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.3.1
packaging==25.0
pdf2image==1.17.0
pillow==11.2.1
pip-review==1.3.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyPDF2==3.0.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
qrcode>=7.4.2,<8.0.0
repath==0.9.0
reportlab==4.4.2
requests==2.32.4
rich==14.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
text-unidecode==1.3
toml==0.10.2
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
watchdog>=4.0.0,<5.0.0
watchfiles==1.1.0
websockets==15.0.1
